backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
elasticsearch
python-dotenv
redis
passlib[bcrypt]
python-jose[cryptography]